Bump react native version for the demo app
