+++ Example/package.json
@@ -7,8 +7,8 @@
     "ios": "node node_modules/react-native/local-cli/cli.js run-ios"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.24.1",
+    "react": "^15.3.1",
+    "react-native": "^0.34.1",
     "react-native-app-intro": "file:.."
   },
   "devDependencies": {
@@ -16,6 +16,7 @@
     "eslint": "^2.8.0",
     "eslint-config-airbnb": "^7.0.0",
     "eslint-plugin-jsx-a11y": "0.6.2",
-    "eslint-plugin-react": "4.3.0"
+    "eslint-plugin-react": "4.3.0",
+    "react-native-cli": "^1.0.0"
   }
 }
